requirements.txt
GitPython>=3.1.40
tree-sitter>=0.25.1
tree-sitter-java>=0.23.5
pytest>=7.0.0